refactor: optional server ssl handler
